fashion_agent/requirements.txt
#mcp
google-generativeai
